+++ awc/src/responses/response_body.rs
@@ -8,22 +8,35 @@ use std::{
 use actix_http::{error::PayloadError, header, HttpMessage};
 use bytes::Bytes;
 use futures_core::Stream;
+use pin_project_lite::pin_project;
 
 use super::{read_body::ReadBody, ResponseTimeout, DEFAULT_BODY_LIMIT};
 use crate::ClientResponse;
 
-/// Future that resolves to a complete response body.
-pub struct ResponseBody<S> {
-    body: Result<ReadBody<S>, Option<PayloadError>>,
-    length: Option<usize>,
-    timeout: ResponseTimeout,
+pin_project! {
+    /// A `Future` that reads a body stream, resolving as [`Bytes`].
+    ///
+    /// # Errors
+    /// `Future` implementation returns error if:
+    /// - content type is not `application/json`;
+    /// - content length is greater than [limit](JsonBody::limit) (default: 2 MiB).
+    pub struct ResponseBody<S> {
+        #[pin]
+        body: Option<ReadBody<S>>,
+        length: Option<usize>,
+        timeout: ResponseTimeout,
+        err: Option<PayloadError>,
+    }
 }
+
+#[deprecated(since = "3.0.0", note = "Renamed to `ResponseBody`.")]
+pub type MessageBody<B> = ResponseBody<B>;
 
 impl<S> ResponseBody<S>
 where
     S: Stream<Item = Result<Bytes, PayloadError>>,
 {
-    /// Create `MessageBody` for request.
+    /// Creates a body stream reader from a response by taking its payload.
     pub fn new(res: &mut ClientResponse<S>) -> ResponseBody<S> {
         let length = match res.headers().get(&header::CONTENT_LENGTH) {
             Some(value) => {
@@ -38,9 +51,10 @@ where
         };
 
         ResponseBody {
+            body: Some(ReadBody::new(res.take_payload(), DEFAULT_BODY_LIMIT)),
             length,
             timeout: mem::take(&mut res.timeout),
-            body: Ok(ReadBody::new(res.take_payload(), DEFAULT_BODY_LIMIT)),
+            err: None,
         }
     }
 
@@ -48,44 +62,46 @@ where
     ///
     /// The default limit is 2 MiB.
     pub fn limit(mut self, limit: usize) -> Self {
-        if let Ok(ref mut body) = self.body {
+        if let Some(ref mut body) = self.body {
             body.limit = limit;
         }
+
         self
     }
 
-    fn err(e: PayloadError) -> Self {
+    fn err(err: PayloadError) -> Self {
         ResponseBody {
+            body: None,
             length: None,
             timeout: ResponseTimeout::default(),
-            body: Err(Some(e)),
+            err: Some(err),
         }
     }
 }
 
 impl<S> Future for ResponseBody<S>
 where
-    S: Stream<Item = Result<Bytes, PayloadError>> + Unpin,
+    S: Stream<Item = Result<Bytes, PayloadError>>,
 {
     type Output = Result<Bytes, PayloadError>;
 
     fn poll(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
-        let this = self.get_mut();
+        let this = self.project();
 
-        match this.body {
-            Err(ref mut err) => Poll::Ready(Err(err.take().unwrap())),
-            Ok(ref mut body) => {
-                if let Some(len) = this.length.take() {
-                    if len > body.limit {
-                        return Poll::Ready(Err(PayloadError::Overflow));
-                    }
-                }
+        if let Some(err) = this.err.take() {
+            return Poll::Ready(Err(err));
+        }
 
-                this.timeout.poll_timeout(cx)?;
-
-                Pin::new(body).poll(cx)
+        if let Some(len) = this.length.take() {
+            let body = Option::as_ref(&this.body).unwrap();
+            if len > body.limit {
+                return Poll::Ready(Err(PayloadError::Overflow));
             }
         }
+
+        this.timeout.poll_timeout(cx)?;
+
+        this.body.as_pin_mut().unwrap().poll(cx)
     }
 }
 
@@ -99,7 +115,7 @@ mod tests {
     assert_impl_all!(ResponseBody<()>: Unpin);
 
     #[actix_rt::test]
-    async fn test_body() {
+    async fn read_body() {
         let mut req = TestResponse::with_header((header::CONTENT_LENGTH, "xxxx")).finish();
         match req.body().await.err().unwrap() {
             PayloadError::UnknownLength => {}